Restructure command files
