+++ commands/loot/index.js
