+++ test/goats.js
@@ -1,11 +1,11 @@
 const assert = require('assert');
-const { validateSpawn } = require('../commands/timer/goats');
+const { validateSpawn } = require('../commands/goats');
 
-describe('goats', function() {
-  describe('validateSpawn', function() {
-    it('should return default nextSpawn and countdown values when in normal flow and updated sheet', function() {
+describe('goats', function () {
+  describe('validateSpawn', function () {
+    it('should return default nextSpawn and countdown values when in normal flow and updated sheet', function () {
       const normalWorldBossData = {
-        nextSpawn: '3:45:00 PM'
+        nextSpawn: '3:45:00 PM',
       };
       const serverTime = 'August 21, 2019 2:00:45 PM';
 
@@ -15,9 +15,9 @@ describe('goats', function() {
       assert.equal(result.countdown, '1 hr 44 mins 15 secs');
     });
 
-    it('should return correct nextSpawn and countdown values when in late night flow and updated sheet', function() {
+    it('should return correct nextSpawn and countdown values when in late night flow and updated sheet', function () {
       const lateNightWorldBossData = {
-        nextSpawn: '1:45:00 AM'
+        nextSpawn: '1:45:00 AM',
       };
       const serverTime = 'August 21, 2019 11:00:45 PM';
 
@@ -27,9 +27,9 @@ describe('goats', function() {
       assert.equal(result.countdown, '2 hrs 44 mins 15 secs');
     });
 
-    it('should return correct nextSpawn and countdown values when in normal flow and no editor updated sheet', function() {
+    it('should return correct nextSpawn and countdown values when in normal flow and no editor updated sheet', function () {
       const normalNonUpdatedWorldBossData = {
-        nextSpawn: '3:45:00 PM'
+        nextSpawn: '3:45:00 PM',
       };
       const serverTime = 'August 21, 2019 5:00:45 PM';
 
@@ -39,9 +39,9 @@ describe('goats', function() {
       assert.equal(result.countdown, '2 hrs 44 mins 15 secs');
     });
 
-    it('should return correct nextSpawn and countdown values when in late night flow and no editor updated sheet', function() {
+    it('should return correct nextSpawn and countdown values when in late night flow and no editor updated sheet', function () {
       const lateNightNonUpdatedWorldBossData = {
-        nextSpawn: '9:45:00 PM'
+        nextSpawn: '9:45:00 PM',
       };
       const serverTime = 'August 21, 2019 11:00:45 PM';
 
@@ -51,9 +51,9 @@ describe('goats', function() {
       assert.equal(result.countdown, '2 hrs 44 mins 15 secs');
     });
 
-    it('should return correct nextSpawn and countdown values when in late night flow and no editor updated sheet and serverTime has gone to next day', function() {
+    it('should return correct nextSpawn and countdown values when in late night flow and no editor updated sheet and serverTime has gone to next day', function () {
       const midNightNonUpdatedWorldBossData = {
-        nextSpawn: '9:45:00 PM'
+        nextSpawn: '9:45:00 PM',
       };
       const serverTime = 'August 22, 2019 1:00:45 AM';
